requirements.txt
adagio==0.2.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
coreforecast==0.0.10
cycler==0.12.1
Cython==3.0.10
darts==0.30.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.6.0
fugue==0.9.1
greenlet==3.0.3
h11==0.14.0
holidays==0.51
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lightning-utilities==0.11.2
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nfoursid==1.0.1
numba==0.60.0
numpy==1.26.4
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
patsy==0.5.6
pillow==10.3.0
pip==24.3.1
platformdirs==4.2.2
pmdarima==2.0.4
protobuf==5.27.1
psycopg2==2.9.9
pyarrow==16.1.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymongo==4.8.0
pyod==2.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.14.0
setuptools==70.1.0
shap==0.45.1
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlmodel==0.0.19
starlette==0.37.2
statsforecast==1.7.5
statsmodels==0.14.2
sympy==1.12.1
tbats==1.1.3
tbb==2021.13.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.3.1
torchmetrics==1.4.0.post0
tqdm==4.66.4
triad==0.9.7
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
utilsforecast==0.1.11
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
xarray==2024.6.0
xgboost==2.1.0
yarl==1.9.4
